+++ GTE/Mathematics/ApprEllipseByArcs.h
@@ -3,7 +3,7 @@
 // Distributed under the Boost Software License, Version 1.0.
 // https://www.boost.org/LICENSE_1_0.txt
 // https://www.geometrictools.com/License/Boost/LICENSE_1_0.txt
-// Version: 4.0.2019.08.13
+// Version: 4.0.2021.11.11
 
 #pragma once
 
@@ -42,7 +42,7 @@ namespace gte
             return false;
         }
 
-        points.resize(numArcs + 1);
+        points.resize(static_cast<size_t>(numArcs) + 1);
         centers.resize(numArcs);
         radii.resize(numArcs);
