+++ GTE/Mathematics/IndexAttribute.h
@@ -3,7 +3,7 @@
 // Distributed under the Boost Software License, Version 1.0.
 // https://www.boost.org/LICENSE_1_0.txt
 // https://www.geometrictools.com/License/Boost/LICENSE_1_0.txt
-// Version: 4.0.2019.08.13
+// Version: 4.0.2021.11.11
 
 #pragma once
 
@@ -31,7 +31,7 @@ namespace gte
         {
             if (size == sizeof(uint32_t))
             {
-                uint32_t* index = reinterpret_cast<uint32_t*>(source) + 3 * t;
+                uint32_t* index = reinterpret_cast<uint32_t*>(source) + 3 * static_cast<size_t>(t);
                 index[0] = v0;
                 index[1] = v1;
                 index[2] = v2;
@@ -40,7 +40,7 @@ namespace gte
 
             if (size == sizeof(uint16_t))
             {
-                uint16_t* index = reinterpret_cast<uint16_t*>(source) + 3 * t;
+                uint16_t* index = reinterpret_cast<uint16_t*>(source) + 3 * static_cast<size_t>(t);
                 index[0] = static_cast<uint16_t>(v0);
                 index[1] = static_cast<uint16_t>(v1);
                 index[2] = static_cast<uint16_t>(v2);
@@ -54,7 +54,7 @@ namespace gte
         {
             if (size == sizeof(uint32_t))
             {
-                uint32_t* index = reinterpret_cast<uint32_t*>(source) + 3 * t;
+                uint32_t* index = reinterpret_cast<uint32_t*>(source) + 3 * static_cast<size_t>(t);
                 v0 = index[0];
                 v1 = index[1];
                 v2 = index[2];
@@ -63,7 +63,7 @@ namespace gte
 
             if (size == sizeof(uint16_t))
             {
-                uint16_t* index = reinterpret_cast<uint16_t*>(source) + 3 * t;
+                uint16_t* index = reinterpret_cast<uint16_t*>(source) + 3 * static_cast<size_t>(t);
                 v0 = static_cast<uint32_t>(index[0]);
                 v1 = static_cast<uint32_t>(index[1]);
                 v2 = static_cast<uint32_t>(index[2]);
