+++ GTE/Mathematics/IntpAkimaNonuniform1.h
@@ -3,7 +3,7 @@
 // Distributed under the Boost Software License, Version 1.0.
 // https://www.boost.org/LICENSE_1_0.txt
 // https://www.geometrictools.com/License/Boost/LICENSE_1_0.txt
-// Version: 4.0.2019.08.13
+// Version: 4.0.2021.11.11
 
 #pragma once
 
@@ -30,7 +30,7 @@ namespace gte
             }
 
             // Compute slopes.
-            std::vector<Real> slope(quantity + 3);
+            std::vector<Real> slope(static_cast<size_t>(quantity) + 3);
             int i, ip1, ip2;
             for (i = 0, ip1 = 1, ip2 = 2; i < quantity - 1; ++i, ++ip1, ++ip2)
             {
@@ -41,8 +41,8 @@ namespace gte
 
             slope[1] = (Real)2 * slope[2] - slope[3];
             slope[0] = (Real)2 * slope[1] - slope[2];
-            slope[quantity + 1] = (Real)2 * slope[quantity] - slope[quantity - 1];
-            slope[quantity + 2] = (Real)2 * slope[quantity + 1] - slope[quantity];
+            slope[static_cast<size_t>(quantity) + 1] = (Real)2 * slope[quantity] - slope[static_cast<size_t>(quantity) - 1];
+            slope[static_cast<size_t>(quantity) + 2] = (Real)2 * slope[static_cast<size_t>(quantity) + 1] - slope[quantity];
 
             // Construct derivatives.
             std::vector<Real> FDer(quantity);
