+++ GTE/Mathematics/LDLTDecomposition.h
@@ -3,7 +3,7 @@
 // Distributed under the Boost Software License, Version 1.0.
 // https://www.boost.org/LICENSE_1_0.txt
 // https://www.geometrictools.com/License/Boost/LICENSE_1_0.txt
-// Version: 5.9.2021.05.13
+// Version: 5.9.2021.11.11
 
 #pragma once
 
@@ -481,7 +481,7 @@ namespace gte
             if (verifySize)
             {
                 LogAssert(
-                    M.size() == NumBlocks * NumBlocks,
+                    M.size() == static_cast<size_t>(NumBlocks) * static_cast<size_t>(NumBlocks),
                     "Invalid size.");
             }
 
@@ -507,7 +507,7 @@ namespace gte
             if (verifySize)
             {
                 LogAssert(
-                    M.size() == NumBlocks * NumBlocks,
+                    M.size() == static_cast<size_t>(NumBlocks) * static_cast<size_t>(NumBlocks),
                     "Invalid size.");
             }
 
@@ -586,7 +586,7 @@ namespace gte
             if (verifySize)
             {
                 LogAssert(
-                    MBlock.size() == NumBlocks * NumBlocks,
+                    MBlock.size() == static_cast<size_t>(NumBlocks) * static_cast<size_t>(NumBlocks),
                     "Invalid size.");
 
                 for (auto const& current : MBlock)
@@ -810,7 +810,7 @@ namespace gte
         // 2-dimensional BlockMatrix object.
         inline size_t GetIndex(int32_t row, int32_t col) const
         {
-            return static_cast<size_t>(col + row * NumBlocks);
+            return static_cast<size_t>(col + row * static_cast<size_t>(NumBlocks));
         }
     };
 }
