+++ GTE/Mathematics/NURBSSurface.h
@@ -3,7 +3,7 @@
 // Distributed under the Boost Software License, Version 1.0.
 // https://www.boost.org/LICENSE_1_0.txt
 // https://www.geometrictools.com/License/Boost/LICENSE_1_0.txt
-// Version: 4.0.2019.12.28
+// Version: 4.0.2021.11.11
 
 #pragma once
 
@@ -104,7 +104,7 @@ namespace gte
         {
             if (0 <= i0 && i0 < GetNumControls(0) && 0 <= i1 && i1 < GetNumControls(1))
             {
-                mControls[i0 + mNumControls[0] * i1] = control;
+                mControls[i0 + static_cast<size_t>(mNumControls[0]) * i1] = control;
             }
         }
 
@@ -112,7 +112,7 @@ namespace gte
         {
             if (0 <= i0 && i0 < GetNumControls(0) && 0 <= i1 && i1 < GetNumControls(1))
             {
-                return mControls[i0 + mNumControls[0] * i1];
+                return mControls[i0 + static_cast<size_t>(mNumControls[0]) * i1];
             }
             else
             {
@@ -124,7 +124,7 @@ namespace gte
         {
             if (0 <= i0 && i0 < GetNumControls(0) && 0 <= i1 && i1 < GetNumControls(1))
             {
-                mWeights[i0 + mNumControls[0] * i1] = weight;
+                mWeights[i0 + static_cast<size_t>(mNumControls[0]) * i1] = weight;
             }
         }
 
@@ -132,7 +132,7 @@ namespace gte
         {
             if (0 <= i0 && i0 < GetNumControls(0) && 0 <= i1 && i1 < GetNumControls(1))
             {
-                return mWeights[i0 + mNumControls[0] * i1];
+                return mWeights[i0 + static_cast<size_t>(mNumControls[0]) * i1];
             }
             else
             {
